+++ docs/html/search/functions_3.js
@@ -2,6 +2,6 @@ var searchData=
 [
   ['enable_5fmulti_5fthread_5for_5fprocess_5fmode',['enable_multi_thread_or_process_mode',['../class_utilities.html#a384728ae9b72982c1351d881ba7ff007',1,'Utilities']]],
   ['err',['err',['../class_utilities.html#a5308da50597c93aad2b1ca2ecfbcb723',1,'Utilities']]],
-  ['exit_5fevent',['exit_event',['../dr__external__ss__events_8cpp.html#a655c944988bd65e5216f857f33e2f174',1,'dr_external_ss_events.cpp']]],
-  ['external_5fevent_5fapp_5finstruction',['external_event_app_instruction',['../dr__external__ss__events_8cpp.html#a14668a99e9c1c591ee2ae89b26748bdf',1,'dr_external_ss_events.cpp']]]
+  ['event_5fapp_5finstruction',['event_app_instruction',['../dr__shadow__stack__client_8cpp.html#a0aaea6e47ff2dc8fb1587381292ba5b5',1,'dr_shadow_stack_client.cpp']]],
+  ['exit_5fevent',['exit_event',['../dr__shadow__stack__client_8cpp.html#ac27bdb8c91453a10aea3bc778eb98f19',1,'dr_shadow_stack_client.cpp']]]
 ];