+++ docs/html/search/functions_4.js
@@ -2,6 +2,5 @@ var searchData=
 [
   ['finish',['finish',['../struct_sym.html#a474eb27bfe01436357e0be478020611c',1,'Sym']]],
   ['flagsforfile',['FlagsForFile',['../_8ycm__extra__conf_8py.html#a51f8bcdc9a3b791e6a88d798e6c786b3',1,'.ycm_extra_conf.py']]],
-  ['fork_5fevent',['fork_event',['../dr__external__ss__events_8cpp.html#ac2cfa3642a80b1480f624e8ed7b36b5c',1,'dr_external_ss_events.cpp']]],
   ['fork_5fhandler',['fork_handler',['../external__stack__server_8cpp.html#a982a481632c7cb255658096b753e3d65',1,'external_stack_server.cpp']]]
 ];