+++ docs/html/search/functions_a.js
@@ -1,6 +1,7 @@
 var searchData=
 [
-  ['on_5fcall',['on_call',['../dr__external__ss__events_8cpp.html#a8161a26ec065b8c83d2f42637896162c',1,'on_call(const app_pc ret_to_addr):&#160;dr_external_ss_events.cpp'],['../dr__internal__ss__events_8cpp.html#a8161a26ec065b8c83d2f42637896162c',1,'on_call(const app_pc ret_to_addr):&#160;dr_internal_ss_events.cpp']]],
-  ['on_5fret',['on_ret',['../dr__external__ss__events_8cpp.html#a42c8b67f8f85fd6bad5bd11892c0d5bc',1,'on_ret(const app_pc instr_addr, const app_pc target_addr):&#160;dr_external_ss_events.cpp'],['../dr__internal__ss__events_8cpp.html#ac3b510b4eb4f19ab3f68d8183f9d802b',1,'on_ret(app_pc, const app_pc target_addr):&#160;dr_internal_ss_events.cpp']]],
+  ['on_5fcall',['on_call',['../dr__external__ss__events_8cpp.html#a8161a26ec065b8c83d2f42637896162c',1,'on_call(const app_pc ret_to_addr):&#160;dr_external_ss_events.cpp'],['../dr__internal__ss__events_8cpp.html#a273951c2f48526aefe80467812c06bf0',1,'on_call(const app_pc ret_to_addr):&#160;dr_internal_ss_events.cpp']]],
+  ['on_5fret',['on_ret',['../dr__external__ss__events_8cpp.html#a42c8b67f8f85fd6bad5bd11892c0d5bc',1,'on_ret(const app_pc instr_addr, const app_pc target_addr):&#160;dr_external_ss_events.cpp'],['../dr__internal__ss__events_8cpp.html#a355e86d607356d27d9e8e6a1b70340db',1,'on_ret(app_pc, const app_pc target_addr):&#160;dr_internal_ss_events.cpp']]],
+  ['on_5fsignal',['on_signal',['../dr__external__ss__events_8cpp.html#ab8a1f864d0acbfeb7d67994f67cbba97',1,'on_signal():&#160;dr_external_ss_events.cpp'],['../dr__internal__ss__events_8cpp.html#a3ab8756aab33706f6a6bf31d0d893add',1,'on_signal():&#160;dr_internal_ss_events.cpp']]],
   ['operator_3d',['operator=',['../class_terminate_on_destruction.html#aef7320686e582846fd4e30c142feec18',1,'TerminateOnDestruction']]]
 ];