+++ docs/html/search/variables_b.js
@@ -1,10 +1,5 @@
 var searchData=
 [
-  ['setup',['setup',['../struct_sym.html#a5b0d02c05e1a5450e6dc79903615ab46',1,'Sym::setup()'],['../class_proc_r_c.html#acccb74d8597ab4b0041ba32ea64ea06d',1,'ProcRC::setup()']]],
-  ['setup_5fcomplete',['setup_complete',['../struct_group.html#a35632ca452cfe09fbf723539cd264f58',1,'Group']]],
-  ['shadow_5fstack',['shadow_stack',['../dr__internal__ss__events_8cpp.html#a4d533d2157d896693f65cc71f4b3bf5e',1,'dr_internal_ss_events.cpp']]],
-  ['size',['size',['../struct_message_1_1_message_type_3_01true_00_01_info_01_4.html#a2cc4f5dbfaba4502c6209e73836c96f1',1,'Message::MessageType&lt; true, Info &gt;::size()'],['../struct_message_1_1_message_type_3_01false_00_01_info_01_4.html#ab1478611666e0b38703dbb0512262d32',1,'Message::MessageType&lt; false, Info &gt;::size()']]],
-  ['sock',['sock',['../dr__external__ss__events_8cpp.html#a5903d0b282fc5eae503de618f896b5e1',1,'dr_external_ss_events.cpp']]],
-  ['source_5fextensions',['SOURCE_EXTENSIONS',['../_8ycm__extra__conf_8py.html#a47014996e1e517071cd0412a22adb123',1,'.ycm_extra_conf.py']]],
-  ['stdout_5ffile',['stdout_file',['../class_utilities.html#a224eea0a234ac9877a8334f74fc89ef0',1,'Utilities']]]
+  ['target',['target',['../struct_args.html#a49f5255b7466ee4e1329aa32ca069da6',1,'Args']]],
+  ['target_5fargs',['target_args',['../struct_args.html#aa9f8e387df0b3691352daf352876176d',1,'Args']]]
 ];